# openapi.json
{"openapi":"3.1.0","info":{"title":"Kiosk API","version":"0.1.0"},"servers":[{"url":"/app"}],"paths":{"/procedures/":{"get":{"tags":["Procedures"],"summary":"List Procedures","operationId":"list_procedures_procedures__get","parameters":[{"required":false,"schema":{"type":"string","title":"Search","default":""},"name":"search","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Procedure"},"type":"array","title":"Response List Procedures Procedures  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/procedures/search-extended":{"get":{"tags":["Procedures"],"summary":"Search Procedures With Counters","operationId":"search_procedures_with_counters_procedures_search_extended_get","parameters":[{"required":false,"schema":{"type":"string","title":"Search","default":""},"name":"search","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProcedureSearchResponse"},"type":"array","title":"Response Search Procedures With Counters Procedures Search Extended Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tickets/":{"post":{"tags":["Tickets"],"summary":"Create Ticket","operationId":"create_ticket_tickets__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seats/":{"get":{"tags":["Seats"],"summary":"List Seats","operationId":"list_seats_seats__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Seat"},"type":"array","title":"Response List Seats Seats  Get"}}}}}}},"/seats/{seat_id}":{"put":{"tags":["Seats"],"summary":"Update Seat","operationId":"update_seat_seats__seat_id__put","parameters":[{"required":true,"schema":{"type":"integer","title":"Seat Id"},"name":"seat_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeatUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Seat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/counters/{counter_id}/call-next":{"post":{"tags":["Counters"],"summary":"Call Next Manually","operationId":"call_next_manually_counters__counter_id__call_next_post","parameters":[{"required":true,"schema":{"type":"integer","title":"Counter Id"},"name":"counter_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalledTicket"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/counters/{counter_id}/pause":{"post":{"tags":["Counters"],"summary":"Pause Counter","operationId":"pause_counter_counters__counter_id__pause_post","parameters":[{"required":true,"schema":{"type":"integer","title":"Counter Id"},"name":"counter_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CounterPauseCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CounterPauseLog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/counters/{counter_id}/resume":{"put":{"tags":["Counters"],"summary":"Resume Counter Route","operationId":"resume_counter_route_counters__counter_id__resume_put","parameters":[{"required":true,"schema":{"type":"integer","title":"Counter Id"},"name":"counter_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Counter"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auths/auth/token":{"post":{"tags":["Authentication"],"summary":"Login For Access Token","operationId":"login_for_access_token_auths_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_auths_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auths/users/":{"post":{"tags":["Authentication"],"summary":"Create User","operationId":"create_user_auths_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auths/me":{"get":{"tags":["Authentication"],"summary":"Get Current User Info","operationId":"get_current_user_info_auths_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_auths_auth_token_post":{"properties":{"grant_type":{"type":"string","pattern":"^password$","title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"type":"string","title":"Client Id"},"client_secret":{"type":"string","format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_auths_auth_token_post"},"CalledTicket":{"properties":{"number":{"type":"integer","title":"Number"},"counter_name":{"type":"string","title":"Counter Name"}},"type":"object","required":["number","counter_name"],"title":"CalledTicket"},"Counter":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"status":{"type":"string","title":"Status","default":"active"}},"type":"object","required":["id","name"],"title":"Counter"},"CounterPauseCreate":{"properties":{"reason":{"type":"string","title":"Reason"}},"type":"object","required":["reason"],"title":"CounterPauseCreate"},"CounterPauseLog":{"properties":{"id":{"type":"integer","title":"Id"},"counter_id":{"type":"integer","title":"Counter Id"},"reason":{"type":"string","title":"Reason"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","counter_id","reason","created_at"],"title":"CounterPauseLog"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Procedure":{"properties":{"name":{"type":"string","title":"Name"},"field_id":{"type":"integer","title":"Field Id"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","field_id","id"],"title":"Procedure"},"ProcedureSearchResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"field_id":{"type":"integer","title":"Field Id"},"counters":{"items":{"$ref":"#/components/schemas/Counter"},"type":"array","title":"Counters"}},"type":"object","required":["id","name","field_id","counters"],"title":"ProcedureSearchResponse"},"Role":{"type":"string","enum":["admin","leader","officer"],"title":"Role","description":"An enumeration."},"Seat":{"properties":{"name":{"type":"string","title":"Name"},"type":{"$ref":"#/components/schemas/SeatType"},"counter_id":{"type":"integer","title":"Counter Id"},"occupied":{"type":"boolean","title":"Occupied","default":false},"id":{"type":"integer","title":"Id"},"last_empty_time":{"type":"string","format":"date-time","title":"Last Empty Time"}},"type":"object","required":["name","type","counter_id","id"],"title":"Seat"},"SeatType":{"type":"string","enum":["officer","client"],"title":"SeatType","description":"An enumeration."},"SeatUpdate":{"properties":{"status":{"type":"boolean","title":"Status"}},"type":"object","required":["status"],"title":"SeatUpdate"},"Ticket":{"properties":{"id":{"type":"integer","title":"Id"},"number":{"type":"integer","title":"Number"},"counter_id":{"type":"integer","title":"Counter Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","number","counter_id","created_at","status"],"title":"Ticket"},"TicketCreate":{"properties":{"counter_id":{"type":"integer","title":"Counter Id"}},"type":"object","required":["counter_id"],"title":"TicketCreate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"username":{"type":"string","title":"Username"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/Role"},"counter_id":{"type":"integer","title":"Counter Id"},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["username","role","id","is_active"],"title":"User"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/Role"},"counter_id":{"type":"integer","title":"Counter Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","role","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}